fix: issues with sizefrom and string/byte slices
The "sizefrom" field is not documented but can be used to limit
the size of the slice to a given number of elements. The issue
is this was not being honored for strings and byte slice types due
to missing logic within the Size() and Pack() where there were
optimizations for string and byte slices.